requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.11
django-excel-response==2.0.3
et-xmlfile==1.0.1
idna==2.6
jdcal==1.3
mysqlclient==1.3.12
openpyxl==2.4.11
python-dateutil==2.7.0
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
